# api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Krishna Pharmacy API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /medicines:
    get:
      operationId: listMedicines
      tags: [medicines]
      summary: List medicines
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of medicines
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Medicine"
    post:
      operationId: createMedicine
      tags: [medicines]
      summary: Create medicine (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicineInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medicine"
  /medicines/{id}:
    get:
      operationId: getMedicine
      tags: [medicines]
      summary: Get medicine by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Medicine detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medicine"
    patch:
      operationId: updateMedicine
      tags: [medicines]
      summary: Update medicine (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicineUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medicine"
    delete:
      operationId: deleteMedicine
      tags: [medicines]
      summary: Delete medicine (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get current cart
      responses:
        "200":
          description: Cart items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemInput"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/{medicineId}:
    delete:
      operationId: removeFromCart
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: medicineId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
components:
  schemas:
...
    Medicine:
      type: object
      required: [id, name, brand, category, mrp, discountPrice, stock, prescriptionRequired, imageUrl]
      properties:
        id:
          type: integer
        name:
          type: string
        brand:
          type: string
        saltComposition:
          type: ["string", "null"]
        category:
          type: string
        description:
          type: ["string", "null"]
        dosage:
          type: ["string", "null"]
        sideEffects:
          type: ["string", "null"]
        storageInstructions:
          type: ["string", "null"]
        manufacturer:
          type: ["string", "null"]
        batchNumber:
          type: ["string", "null"]
        expiryDate:
          type: ["string", "null"]
        mrp:
          type: number
        discountPrice:
          type: number
        stock:
          type: integer
        prescriptionRequired:
          type: boolean
        imageUrl:
          type: string
        isFeatured:
          type: boolean
        createdAt:
          type: string
    MedicineInput:
      type: object
      required: [name, brand, category, mrp, discountPrice, stock, prescriptionRequired]
      properties:
        name:
          type: string
        brand:
          type: string
        saltComposition:
          type: string
        category:
          type: string
        description:
          type: string
        dosage:
          type: string
        sideEffects:
          type: string
        storageInstructions:
          type: string
        manufacturer:
          type: string
        batchNumber:
          type: string
        expiryDate:
          type: string
        mrp:
          type: number
        discountPrice:
          type: number
        stock:
          type: integer
        prescriptionRequired:
          type: boolean
        imageUrl:
          type: string
        isFeatured:
          type: boolean
    MedicineUpdate:
      type: object
      properties:
        name:
          type: string
        brand:
          type: string
        saltComposition:
          type: string
        category:
          type: string
        description:
          type: string
        dosage:
          type: string
        sideEffects:
          type: string
        storageInstructions:
          type: string
        manufacturer:
          type: string
        batchNumber:
          type: string
        expiryDate:
          type: string
        mrp:
          type: number
        discountPrice:
          type: number
        stock:
          type: integer
        prescriptionRequired:
          type: boolean
        imageUrl:
          type: string
        isFeatured:
          type: boolean
...
    Cart:
      type: object
      required: [items, totalAmount]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        totalAmount:
          type: number
        totalItems:
          type: integer
    CartItem:
      type: object
      required: [medicineId, medicineName, quantity, price, imageUrl]
      properties:
        medicineId:
          type: integer
        medicineName:
          type: string
        quantity:
          type: integer
        price:
          type: number
        imageUrl:
          type: string
        prescriptionRequired:
          type: boolean
    CartItemInput:
      type: object
      required: [medicineId, quantity]
      properties:
        medicineId:
          type: integer
        quantity:
          type: integer
    Order:
      type: object
      required: [id, status, totalAmount, createdAt, items]
      properties:
        id:
          type: integer
        customerName:
          type: ["string", "null"]
        customerPhone:
          type: ["string", "null"]
        deliveryAddress:
          type: ["string", "null"]
        status:
          type: string
        totalAmount:
          type: number
        prescriptionId:
          type: ["integer", "null"]
        paymentStatus:
          type: string
        createdAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
    OrderItem:
      type: object
      required: [medicineId, medicineName, quantity, price]
      properties:
        medicineId:
          type: integer
        medicineName:
          type: string
        quantity:
          type: integer
        price:
          type: number
    OrderInput:
      type: object
      required: [items, deliveryAddress, customerName, customerPhone]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItemInput"
        deliveryAddress:
          type: string
        customerName:
          type: string
        customerPhone:
          type: string
        prescriptionId:
          type: integer
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
...