# src/api-zod/generated/types/cart.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Krishna Pharmacy API
 * OpenAPI spec version: 0.1.0
 */
import type { CartItem } from './cartItem';

export interface Cart {
  items: CartItem[];
  totalAmount: number;
  totalItems?: number;
}