# src/api-zod/generated/types/medicineUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Krishna Pharmacy API
 * OpenAPI spec version: 0.1.0
 */

export interface MedicineUpdate {
  name?: string;
  brand?: string;
  saltComposition?: string;
  category?: string;
  description?: string;
  dosage?: string;
  sideEffects?: string;
  storageInstructions?: string;
  manufacturer?: string;
  batchNumber?: string;
  expiryDate?: string;
  mrp?: number;
  discountPrice?: number;
  stock?: number;
  prescriptionRequired?: boolean;
  imageUrl?: string;
  isFeatured?: boolean;
}