# src/api-zod/generated/types/order.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Krishna Pharmacy API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';

export interface Order {
  id: number;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  customerPhone?: string | null;
  /** @nullable */
  deliveryAddress?: string | null;
  status: string;
  totalAmount: number;
  /** @nullable */
  prescriptionId?: number | null;
  paymentStatus?: string;
  createdAt: string;
  items: OrderItem[];
}